requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
langgraph>=0.2.0
openai>=1.58.0
python-dotenv>=1.0.0

